+++ app_flutter/lib/build_dashboard.dart
@@ -3,10 +3,9 @@
 // found in the LICENSE file.
 
 import 'package:flutter/material.dart';
-import 'package:google_sign_in/widgets.dart';
 import 'package:provider/provider.dart';
 
-import 'service/google_authentication.dart';
+import 'sign_in_button.dart';
 import 'state/flutter_build.dart';
 import 'status_grid.dart';
 
@@ -55,7 +54,7 @@ class BuildDashboard extends StatelessWidget {
               ? theme.primaryColor
               : theme.errorColor,
           actions: <Widget>[
-            UserAvatar(buildState: buildState),
+            SignInButton(buildState: buildState),
           ],
         ),
         body: Column(
@@ -64,39 +63,6 @@ class BuildDashboard extends StatelessWidget {
           ],
         ),
       ),
-    );
-  }
-}
-
-/// Widget for displaying sign in information for the current user.
-///
-/// If logged in, it will display the user's avatar. Otherwise, it will show
-/// a button for sign in.
-class UserAvatar extends StatelessWidget {
-  const UserAvatar({@required this.buildState, Key key}) : super(key: key);
-
-  final FlutterBuildState buildState;
-
-  @override
-  Widget build(BuildContext context) {
-    final GoogleSignInService authService = buildState.authService;
-
-    if (authService.isAuthenticated) {
-      return SizedBox(
-        width: 50,
-        height: 50,
-        child: GoogleUserCircleAvatar(
-          identity: authService.user,
-        ),
-      );
-    }
-
-    return FlatButton(
-      child: const Text(
-        'Sign in',
-        style: TextStyle(color: Colors.white),
-      ),
-      onPressed: () => buildState.signIn(),
     );
   }
 }
